[test] Extract REST catalog metadata handlers
